AddressInfo structure.

+++ src/main/scala/org/ergoplatform/explorer/http/api/models/Items.scala
@@ -3,7 +3,7 @@ package org.ergoplatform.explorer.http.api.models
 import io.circe.Codec
 import io.circe.generic.semiauto.deriveCodec
 
-final case class Items[A](items: List[A], total: Long)
+final case class Items[A](items: List[A], total: Int)
 
 object Items {
